Bad__0.1
This version of the duckres.py includes three functions: the first one is similar to that in the master branch, the second one prints the first results, and the third one which automates the process of cropping the image.
The last function has to be fixed, so wait or help.